My code is going back and forth between an NOT NULL error and a UNIQUE constraint error. The id column is set as an autoincrementedPrimaryKey, which I assume has a NOT NULL constraint by default. In the documentation here: https://swiftpackageindex.com/groue/grdb.swift/v6.15.1/documentation/grdb/recordrecommendedpractices
We see that we can set the id's type to Int64?, an optional. Later, we see that we can initialize an object without explicitly setting the id value. We also see that we can make it so that a record learns about its id after insertion. When I do all that, I'm getting a NOT NULL error.

Alternatively, If I set a value for id when I initialize, I'm getting a UNIQUE constraint error.